+++ packages/google_mobile_ads/ios/Tests/FLTGoogleMobileAdsReaderWriterTest.m
@@ -49,6 +49,89 @@
   XCTAssertEqualObjects(decodedSize.height, @(2));
 }
 
+- (void)testEncodeDecodeInlineAdaptiveBannerAdSize_currentOrientation {
+  GADAdSize testAdSize = GADAdSizeFromCGSize(CGSizeMake(25, 10));
+
+  FLTAdSizeFactory *factory = OCMClassMock([FLTAdSizeFactory class]);
+  OCMStub([factory currentOrientationInlineAdaptiveBannerSizeWithWidth:@(23)])
+      .andReturn(testAdSize);
+
+  FLTInlineAdaptiveBannerSize *inlineAdaptiveBannerSize =
+      [[FLTInlineAdaptiveBannerSize alloc] initWithFactory:factory
+                                                     width:@(23)
+                                                 maxHeight:NULL
+                                               orientation:NULL];
+
+  NSData *encodedMessage = [_messageCodec encode:inlineAdaptiveBannerSize];
+
+  FLTInlineAdaptiveBannerSize *decodedSize = [_messageCodec decode:encodedMessage];
+  XCTAssertEqual(decodedSize.size.size.width, testAdSize.size.width);
+  XCTAssertEqualObjects(decodedSize.maxHeight, inlineAdaptiveBannerSize.maxHeight);
+  XCTAssertEqualObjects(decodedSize.orientation, inlineAdaptiveBannerSize.orientation);
+}
+
+- (void)testEncodeDecodeInlineAdaptiveBannerAdSize_portraitOrientation {
+  GADAdSize testAdSize = GADAdSizeFromCGSize(CGSizeMake(25, 10));
+
+  FLTAdSizeFactory *factory = OCMClassMock([FLTAdSizeFactory class]);
+  OCMStub([factory portraitOrientationInlineAdaptiveBannerSizeWithWidth:@(23)])
+      .andReturn(testAdSize);
+
+  FLTInlineAdaptiveBannerSize *inlineAdaptiveBannerSize =
+      [[FLTInlineAdaptiveBannerSize alloc] initWithFactory:factory
+                                                     width:@(23)
+                                                 maxHeight:NULL
+                                               orientation:@0];
+
+  NSData *encodedMessage = [_messageCodec encode:inlineAdaptiveBannerSize];
+
+  FLTInlineAdaptiveBannerSize *decodedSize = [_messageCodec decode:encodedMessage];
+  XCTAssertEqual(decodedSize.size.size.width, testAdSize.size.width);
+  XCTAssertEqualObjects(decodedSize.maxHeight, inlineAdaptiveBannerSize.maxHeight);
+  XCTAssertEqualObjects(decodedSize.orientation, inlineAdaptiveBannerSize.orientation);
+}
+
+- (void)testEncodeDecodeInlineAdaptiveBannerAdSize_landscapeOrientation {
+  GADAdSize testAdSize = GADAdSizeFromCGSize(CGSizeMake(25, 10));
+
+  FLTAdSizeFactory *factory = OCMClassMock([FLTAdSizeFactory class]);
+  OCMStub([factory landscapeInlineAdaptiveBannerAdSizeWithWidth:@(23)]).andReturn(testAdSize);
+
+  FLTInlineAdaptiveBannerSize *inlineAdaptiveBannerSize =
+      [[FLTInlineAdaptiveBannerSize alloc] initWithFactory:factory
+                                                     width:@(23)
+                                                 maxHeight:NULL
+                                               orientation:@1];
+
+  NSData *encodedMessage = [_messageCodec encode:inlineAdaptiveBannerSize];
+
+  FLTInlineAdaptiveBannerSize *decodedSize = [_messageCodec decode:encodedMessage];
+  XCTAssertEqual(decodedSize.size.size.width, testAdSize.size.width);
+  XCTAssertEqualObjects(decodedSize.maxHeight, inlineAdaptiveBannerSize.maxHeight);
+  XCTAssertEqualObjects(decodedSize.orientation, inlineAdaptiveBannerSize.orientation);
+}
+
+- (void)testEncodeDecodeInlineAdaptiveBannerAdSize_withMaxHeight {
+  GADAdSize testAdSize = GADAdSizeFromCGSize(CGSizeMake(25, 10));
+
+  FLTAdSizeFactory *factory = OCMClassMock([FLTAdSizeFactory class]);
+  OCMStub([factory inlineAdaptiveBannerAdSizeWithWidthAndMaxHeight:@(23) maxHeight:@50])
+      .andReturn(testAdSize);
+
+  FLTInlineAdaptiveBannerSize *inlineAdaptiveBannerSize =
+      [[FLTInlineAdaptiveBannerSize alloc] initWithFactory:factory
+                                                     width:@(23)
+                                                 maxHeight:@50
+                                               orientation:nil];
+
+  NSData *encodedMessage = [_messageCodec encode:inlineAdaptiveBannerSize];
+
+  FLTInlineAdaptiveBannerSize *decodedSize = [_messageCodec decode:encodedMessage];
+  XCTAssertEqual(decodedSize.size.size.width, testAdSize.size.width);
+  XCTAssertEqualObjects(decodedSize.maxHeight, inlineAdaptiveBannerSize.maxHeight);
+  XCTAssertEqualObjects(decodedSize.orientation, inlineAdaptiveBannerSize.orientation);
+}
+
 - (void)testEncodeDecodeAnchoredAdaptiveBannerAdSize {
   GADAdSize testAdSize = GADAdSizeFromCGSize(CGSizeMake(0, 0));
 
@@ -98,13 +181,19 @@
   request.keywords = @[ @"apple" ];
   request.contentURL = @"banana";
   request.nonPersonalizedAds = YES;
-
+  NSArray<NSString *> *contentURLs = @[ @"url-1.com", @"url-2.com" ];
+  request.neighboringContentURLs = contentURLs;
+  request.location = [[FLTLocationParams alloc] initWithAccuracy:@1.5 longitude:@52 latitude:@123];
   NSData *encodedMessage = [_messageCodec encode:request];
 
   FLTAdRequest *decodedRequest = [_messageCodec decode:encodedMessage];
   XCTAssertTrue([decodedRequest.keywords isEqualToArray:@[ @"apple" ]]);
   XCTAssertEqualObjects(decodedRequest.contentURL, @"banana");
   XCTAssertTrue(decodedRequest.nonPersonalizedAds);
+  XCTAssertEqualObjects(decodedRequest.neighboringContentURLs, contentURLs);
+  XCTAssertEqualObjects(decodedRequest.location.accuracy, @1.5);
+  XCTAssertEqualObjects(decodedRequest.location.longitude, @52);
+  XCTAssertEqualObjects(decodedRequest.location.latitude, @123);
 }
 
 - (void)testEncodeDecodeGAMAdRequest {
@@ -114,6 +203,10 @@
   request.customTargeting = @{@"table" : @"linen"};
   request.customTargetingLists = @{@"go" : @[ @"lakers" ]};
   request.nonPersonalizedAds = YES;
+  NSArray<NSString *> *contentURLs = @[ @"url-1.com", @"url-2.com" ];
+  request.neighboringContentURLs = contentURLs;
+  request.pubProvidedID = @"pub-id";
+  request.location = [[FLTLocationParams alloc] initWithAccuracy:@1.5 longitude:@52 latitude:@123];
   NSData *encodedMessage = [_messageCodec encode:request];
 
   FLTGAMAdRequest *decodedRequest = [_messageCodec decode:encodedMessage];
@@ -123,6 +216,11 @@
   XCTAssertTrue(
       [decodedRequest.customTargetingLists isEqualToDictionary:@{@"go" : @[ @"lakers" ]}]);
   XCTAssertTrue(decodedRequest.nonPersonalizedAds);
+  XCTAssertEqualObjects(decodedRequest.neighboringContentURLs, contentURLs);
+  XCTAssertEqualObjects(decodedRequest.pubProvidedID, @"pub-id");
+  XCTAssertEqualObjects(decodedRequest.location.accuracy, @1.5);
+  XCTAssertEqualObjects(decodedRequest.location.longitude, @52);
+  XCTAssertEqualObjects(decodedRequest.location.latitude, @123);
 }
 
 - (void)testEncodeDecodeRewardItem {