+++ packages/google_mobile_ads/lib/src/ad_containers.dart
@@ -143,6 +143,40 @@ class LoadAdError extends AdError {
   }
 }
 
+/// Location parameters that can be configured in an ad request.
+class LocationParams {
+  /// Location parameters that can be configured in an ad request.
+  const LocationParams({
+    required this.accuracy,
+    required this.longitude,
+    required this.latitude,
+    this.time,
+  });
+
+  /// The accuracy in meters.
+  final double accuracy;
+
+  /// The longitude in degrees.
+  final double longitude;
+
+  /// The latitude in degrees.
+  final double latitude;
+
+  /// The UTC time, in milliseconds since epoch (January 1, 1970).
+  ///
+  /// This is required on Android, and ignored on iOS.
+  final int? time;
+
+  @override
+  bool operator ==(Object other) {
+    return other is LocationParams &&
+        accuracy == other.accuracy &&
+        longitude == other.longitude &&
+        latitude == other.latitude &&
+        time == other.time;
+  }
+}
+
 /// Targeting info per the AdMob API.
 ///
 /// This class's properties mirror the native AdRequest API. See for example:
@@ -152,7 +186,10 @@ class AdRequest {
   const AdRequest({
     this.keywords,
     this.contentUrl,
+    this.neighboringContentUrls,
     this.nonPersonalizedAds,
+    this.httpTimeoutMillis,
+    this.location,
   });
 
   /// Words or phrases describing the current user activity.
@@ -162,6 +199,9 @@ class AdRequest {
   ///
   /// This webpage content is used for targeting and brand safety purposes.
   final String? contentUrl;
+
+  /// URLs representing web content near an ad.
+  final List<String>? neighboringContentUrls;
 
   /// Non-personalized ads are ads that are not based on a user’s past behavior.
   ///
@@ -169,55 +209,71 @@ class AdRequest {
   /// https://support.google.com/admob/answer/7676680?hl=en
   final bool? nonPersonalizedAds;
 
+  /// A custom timeout for HTTPS calls during an ad request.
+  ///
+  /// This is only supported in Android. This value is ignored on iOS.
+  final int? httpTimeoutMillis;
+
+  /// Location data.
+  ///
+  /// Used for mediation targeting purposes.
+  final LocationParams? location;
+
   @override
   bool operator ==(Object other) {
     return other is AdRequest &&
         listEquals<String>(keywords, other.keywords) &&
         contentUrl == other.contentUrl &&
-        nonPersonalizedAds == other.nonPersonalizedAds;
+        nonPersonalizedAds == other.nonPersonalizedAds &&
+        listEquals(neighboringContentUrls, other.neighboringContentUrls) &&
+        httpTimeoutMillis == other.httpTimeoutMillis &&
+        location == other.location;
   }
 }
 
 /// Targeting info per the Ad Manager API.
-class AdManagerAdRequest {
+class AdManagerAdRequest extends AdRequest {
   /// Constructs an [AdManagerAdRequest] from optional targeting information.
   const AdManagerAdRequest({
-    this.keywords,
-    this.contentUrl,
+    List<String>? keywords,
+    String? contentUrl,
+    List<String>? neighboringContentUrls,
     this.customTargeting,
     this.customTargetingLists,
-    this.nonPersonalizedAds,
-  });
-
-  /// Words or phrases describing the current user activity.
-  final List<String>? keywords;
-
-  /// URL string for a webpage whose content matches the app’s primary content.
-  ///
-  /// This webpage content is used for targeting and brand safety purposes.
-  final String? contentUrl;
+    bool? nonPersonalizedAds,
+    int? httpTimeoutMillis,
+    this.publisherProvidedId,
+    LocationParams? location,
+  }) : super(
+          keywords: keywords,
+          contentUrl: contentUrl,
+          neighboringContentUrls: neighboringContentUrls,
+          nonPersonalizedAds: nonPersonalizedAds,
+          httpTimeoutMillis: httpTimeoutMillis,
+          location: location,
+        );
 
   /// Key-value pairs used for custom targeting.
   final Map<String, String>? customTargeting;
 
   /// Key-value pairs used for custom targeting.
+  ///
+  /// Any duplicate keys from [customTargeting] will be overwritten.
   final Map<String, List<String>>? customTargetingLists;
 
-  /// Non-personalized ads are ads that are not based on a user’s past behavior.
-  ///
-  /// For more information:
-  /// https://support.google.com/admanager/answer/9005435?hl=en
-  final bool? nonPersonalizedAds;
+  /// The identifier used for frequency capping, audience segmentation
+  /// and targeting, sequential ad rotation, and other audience-based ad
+  /// delivery controls across devices.
+  final String? publisherProvidedId;
 
   @override
   bool operator ==(Object other) {
-    return other is AdManagerAdRequest &&
-        listEquals<String>(keywords, other.keywords) &&
-        contentUrl == other.contentUrl &&
+    return super == other &&
+        other is AdManagerAdRequest &&
         mapEquals<String, String>(customTargeting, other.customTargeting) &&
         customTargetingLists.toString() ==
             other.customTargetingLists.toString() &&
-        nonPersonalizedAds == other.nonPersonalizedAds;
+        publisherProvidedId == other.publisherProvidedId;
   }
 }
 
@@ -256,6 +312,42 @@ class SmartBannerAdSize extends AdSize {
 class FluidAdSize extends AdSize {
   /// Default constructor for [FluidAdSize].
   const FluidAdSize() : super(width: -3, height: -3);
+}
+
+/// A size for inline adaptive banner ads.
+///
+/// This ad size takes a width and optionally a maximum height and orientation.
+/// They will be rendered at the provided width and have variable height decided
+/// by Google servers.
+///
+/// You can obtain an instance of this using factory methods such as
+/// [AdSize.getCurrentOrientationInlineAdaptiveBannerAdSize(width)].
+///
+/// After an inline adaptive banner ad is loaded, you should use
+/// [BannerAd.getPlatformAdSize] or [AdManagerBannerAd.getPlatformAdSize] to get
+/// the height to use for the ad container.
+class InlineAdaptiveSize extends AdSize {
+  /// Private constructor for [InlineAdaptiveSize].
+  ///
+  /// You should use one of the static constructors in [AdSize].
+  const InlineAdaptiveSize._({
+    required int width,
+    this.maxHeight,
+    this.orientation,
+  }) : super(width: width, height: 0);
+
+  /// The maximum height allowed.
+  final int? maxHeight;
+
+  /// The orientation. If none if specified the current orientation is used.
+  final Orientation? orientation;
+
+  /// Representation of [orientation] to pass to platform code.
+  int? get orientationValue => orientation == null
+      ? null
+      : orientation == Orientation.portrait
+          ? 0
+          : 1;
 }
 
 /// [AdSize] represents the size of a banner ad.
@@ -320,6 +412,65 @@ class AdSize {
       width: width,
       height: height.truncate(),
     );
+  }
+
+  /// Gets an AdSize with the given width and height that is always 0.
+  ///
+  /// This ad size allows Google servers to choose an optimal ad size with a
+  /// height less than or equal to the height of the screen in the current
+  /// orientation.
+  /// After the ad is loaded, you should update the height of the ad container
+  /// by calling [BannerAd.getPlatformAdSize] or
+  /// [AdManagerBannerAd.getPlatformAdSize] from the ad load callback.
+  /// This ad size is most suitable for ads intended to be displayed inside
+  /// scrollable content.
+  static InlineAdaptiveSize getCurrentOrientationInlineAdaptiveBannerAdSize(
+      int width) {
+    return InlineAdaptiveSize._(width: width);
+  }
+
+  /// Gets an AdSize in landscape orientation with the given width and 0 height.
+  ///
+  /// This ad size allows Google servers to choose an optimal ad size with a
+  /// height less than or equal to the height of the screen in landscape
+  /// orientation.
+  /// After the ad is loaded, you should update the height of the ad container
+  /// by calling [BannerAd.getPlatformAdSize] or
+  /// [AdManagerBannerAd.getPlatformAdSize] from the ad load callback.
+  /// This ad size is most suitable for ads intended to be displayed inside
+  /// scrollable content.
+  static InlineAdaptiveSize getLandscapeInlineAdaptiveBannerAdSize(int width) {
+    return InlineAdaptiveSize._(
+        width: width, orientation: Orientation.landscape);
+  }
+
+  /// Gets an AdSize in portrait orientation with the given width and 0 height.
+  ///
+  /// This ad size allows Google servers to choose an optimal ad size with a
+  /// height less than or equal to the height of the screen in portrait
+  /// orientation.
+  /// After the ad is loaded, you should update the height of the ad container
+  /// by calling [BannerAd.getPlatformAdSize] or
+  /// [AdManagerBannerAd.getPlatformAdSize] from the ad load callback.
+  /// This ad size is most suitable for ads intended to be displayed inside
+  /// scrollable content.
+  static InlineAdaptiveSize getPortraitInlineAdaptiveBannerAdSize(int width) {
+    return InlineAdaptiveSize._(
+        width: width, orientation: Orientation.portrait);
+  }
+
+  /// Gets an AdSize with the given width and height that is always 0.
+  ///
+  /// This ad size allows Google servers to choose an optimal ad size with a
+  /// height less than or equal to [maxHeight].
+  /// After the ad is loaded, you should update the height of the ad container
+  /// by calling [BannerAd.getPlatformAdSize] or
+  /// [AdManagerBannerAd.getPlatformAdSize] from the ad load callback.
+  /// This ad size is most suitable for ads intended to be displayed inside
+  /// scrollable content.
+  static InlineAdaptiveSize getInlineAdaptiveBannerAdSize(
+      int width, int maxHeight) {
+    return InlineAdaptiveSize._(width: width, maxHeight: maxHeight);
   }
 
   /// Ad units that render screen-width banner ads on any screen size across different devices in either orientation on Android.
@@ -679,6 +830,15 @@ class BannerAd extends AdWithView {
   Future<void> load() async {
     await instanceManager.loadBannerAd(this);
   }
+
+  /// Returns the AdSize of the associated platform ad object.
+  ///
+  /// The dimensions of the [AdSize] returned here may differ from [size],
+  /// depending on what type of [AdSize] was used.
+  /// The future will resolve to null if [load] has not been called yet.
+  Future<AdSize?> getPlatformAdSize() async {
+    return await instanceManager.getAdSize(this);
+  }
 }
 
 /// An 'AdManagerBannerAd' that has fluid ad size.
@@ -746,6 +906,15 @@ class AdManagerBannerAd extends AdWithView {
   @override
   Future<void> load() async {
     await instanceManager.loadAdManagerBannerAd(this);
+  }
+
+  /// Returns the AdSize of the associated platform ad object.
+  ///
+  /// The dimensions of the [AdSize] returned here may differ from [size],
+  /// depending on what type of [AdSize] was used.
+  /// The future will resolve to null if [load] has not been called yet.
+  Future<AdSize?> getPlatformAdSize() async {
+    return await instanceManager.getAdSize(this);
   }
 }